Extract padel scoring engine to core module
