+++ wear/build.gradle.kts
@@ -56,6 +56,7 @@ dependencies {
     // Data Layer
     implementation(libs.play.services.wearable)
     implementation(libs.kotlinx.coroutines.play.services)
+    implementation(project(":core-scoring"))
 
 
     // Compose BOM
