+++ wear/src/main/java/com/johan/wear/presentation/MatchViewModel.kt
@@ -2,6 +2,8 @@ package com.johan.wear.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.johan.racketmatchapp.core.scoring.padel.GameEvent
+import com.johan.racketmatchapp.core.scoring.padel.PadelEngine
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
@@ -10,133 +12,6 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-
-enum class PadelScore {
-    LOVE, FIFTEEN, THIRTY, FORTY, ADVANTAGE
-}
-
-data class PlayerState(
-    var score: PadelScore = PadelScore.LOVE,
-    var gameScore: Int = 0,
-    var setScore: Int = 0,
-    var TieBreakScore: Int = 0
-)
-
-sealed interface GameEvent {
-    data class GameOver(val player: Int) : GameEvent
-    data class TieBreak(val player: Int) : GameEvent
-    data class Advantage(val player: Int) : GameEvent
-    data class Score(val player: Int) : GameEvent
-    data class UndoScore(val player: Int) : GameEvent
-    data class Deuce(val player: Int) : GameEvent
-    data class PaddelGameWon(val player: Int) : GameEvent
-    data class SetScore(val player: Int) : GameEvent
-}
-
-class PaddelEngine(private val setLimit: Int = 5) {
-    private val players = listOf(PlayerState(), PlayerState())
-    var tieBreak: Boolean = false
-    var inDeuce: Boolean = false
-
-    fun getp1DisplayScore(): String = if (tieBreak) players[0].TieBreakScore.toString() else padelScoreToString(players[0].score)
-    fun getp2DisplayScore(): String = if (tieBreak) players[1].TieBreakScore.toString() else padelScoreToString(players[1].score)
-    fun get1DisplayGame(): String = players[0].gameScore.toString()
-    fun get2DisplayGame(): String = players[1].gameScore.toString()
-    fun get1DisplaySet(): String = players[0].setScore.toString()
-    fun get2DisplaySet(): String = players[1].setScore.toString()
-
-    fun setTieBreakTrue() { tieBreak = true }
-
-    private fun padelScoreToString(score: PadelScore): String = when (score) {
-        PadelScore.LOVE -> "0"
-        PadelScore.FIFTEEN -> "15"
-        PadelScore.THIRTY -> "30"
-        PadelScore.FORTY -> "40"
-        PadelScore.ADVANTAGE -> "ADV"
-    }
-
-    fun increaseScore(scoringPlayer: Boolean): GameEvent = if (tieBreak) tiebreakThingy(scoringPlayer) else calcScore(scoringPlayer)
-
-    private fun tiebreakThingy(isP1: Boolean): GameEvent {
-        val idx = if (isP1) 0 else 1
-        val oppIdx = if (isP1) 1 else 0
-        players[idx].TieBreakScore++
-        if (players[idx].TieBreakScore >= 7 && (players[idx].TieBreakScore - players[oppIdx].TieBreakScore) >= 2) {
-            players[idx].setScore++
-            return if (players[idx].setScore >= (setLimit / 2) + 1) {
-                resetAfterSetWin()
-                GameEvent.GameOver(idx)
-            } else {
-                resetAfterSetWin()
-                GameEvent.SetScore(idx)
-            }
-        }
-        return GameEvent.Score(idx)
-    }
-
-    private fun calcScore(isP1: Boolean): GameEvent {
-        val idx = if (isP1) 0 else 1
-        val oppIdx = if (isP1) 1 else 0
-        if (inDeuce) {
-            if (players[idx].score == PadelScore.FORTY) {
-                if (players[oppIdx].score == PadelScore.ADVANTAGE) {
-                    players[oppIdx].score = PadelScore.FORTY
-                    players[idx].score = PadelScore.FORTY
-                    return GameEvent.Deuce(idx)
-                } else {
-                    players[idx].score = PadelScore.ADVANTAGE
-                    return GameEvent.Advantage(idx)
-                }
-            } else {
-                return gameWin(idx, oppIdx)
-            }
-        } else {
-            val next = simpleIncrease(players[idx].score)
-            if (next == PadelScore.FORTY && players[oppIdx].score == PadelScore.FORTY) {
-                players[idx].score = PadelScore.FORTY
-                inDeuce = true
-                return GameEvent.Deuce(idx)
-            } else if (players[idx].score == PadelScore.FORTY) {
-                return gameWin(idx, oppIdx)
-            } else {
-                players[idx].score = next
-                return GameEvent.Score(idx)
-            }
-        }
-    }
-
-    private fun gameWin(idx: Int, oppIdx: Int): GameEvent {
-        players[idx].gameScore++
-        if (players[idx].gameScore >= 6 && (players[idx].gameScore - players[oppIdx].gameScore) >= 2) {
-            players[idx].setScore++
-            resetAfterSetWin()
-            if (players[idx].setScore >= (setLimit / 2) + 1) return GameEvent.GameOver(idx)
-        } else if (players[idx].gameScore == 6 && players[oppIdx].gameScore == 6) {
-            resetAfterGameWin()
-            tieBreak = true
-            return GameEvent.TieBreak(idx)
-        }
-        resetAfterGameWin()
-        return GameEvent.PaddelGameWon(idx)
-    }
-
-    private fun resetAfterGameWin() {
-        players.forEach { it.score = PadelScore.LOVE; it.TieBreakScore = 0 }
-        inDeuce = false
-    }
-
-    private fun resetAfterSetWin() {
-        players.forEach { it.score = PadelScore.LOVE; it.gameScore = 0; it.TieBreakScore = 0 }
-        tieBreak = false
-    }
-
-    private fun simpleIncrease(score: PadelScore): PadelScore = when (score) {
-        PadelScore.LOVE -> PadelScore.FIFTEEN
-        PadelScore.FIFTEEN -> PadelScore.THIRTY
-        PadelScore.THIRTY -> PadelScore.FORTY
-        else -> PadelScore.ADVANTAGE
-    }
-}
 
 data class MatchUiState(
     val p1Score: String = "0",
@@ -149,7 +24,7 @@ data class MatchUiState(
 )
 
 class MatchViewModel : ViewModel() {
-    private var engine = PaddelEngine(setLimit = 3)
+    private var engine = PadelEngine(setLimit = 3)
     private val scoreHistory = mutableListOf<Boolean>() // true = P1, false = P2
     
     private val _uiState = MutableStateFlow(MatchUiState())
@@ -172,8 +47,13 @@ class MatchViewModel : ViewModel() {
         if (scoreHistory.isNotEmpty()) {
             scoreHistory.removeAt(scoreHistory.size - 1)
             // Re-calculate state from scratch
-            engine = PaddelEngine(setLimit = 3)
-            scoreHistory.forEach { engine.increaseScore(it) }
+            engine = PadelEngine(setLimit = 3)
+            scoreHistory.forEach { scoredByP1 ->
+                val event = engine.increaseScore(scoredByP1)
+                if (event is GameEvent.TieBreak) {
+                    engine.setTieBreakTrue()
+                }
+            }
             updateUI()
             viewModelScope.launch { _events.emit(GameEvent.UndoScore(0)) }
         }
@@ -181,6 +61,9 @@ class MatchViewModel : ViewModel() {
 
     private fun applyScore(block: () -> GameEvent) {
         val event = block()
+        if (event is GameEvent.TieBreak) {
+            engine.setTieBreakTrue()
+        }
         updateUI()
         viewModelScope.launch { _events.emit(event) }
     }
